[Silver I] Title: 카드 합체 놀이, Time: 0 ms, Memory: 2160 KB -BaekjoonHub

+++ 백준/Silver/15903. 카드 합체 놀이/카드 합체 놀이.cc
@@ -16,7 +16,7 @@ int main(){
     cin >> n >> m;
 
     for(int i=0; i<n; i++){
-        long long int num;
+        int num;
         cin >> num;
         pq.push(num);
     }
